# apps/remote-server/suncode-web/src/main/java/ai/suncode/controller/MobileRemoteController.java
package ai.suncode.controller;

import ai.suncode.common.exception.BusinessException;
import ai.suncode.common.utils.MarshallingUtils;
import ai.suncode.message.ApiBaseRet;
import ai.suncode.message.remote.*;
import ai.suncode.common.http.ServiceContext;
import ai.suncode.service.RemoteAuthService;
import ai.suncode.service.RemoteRelayService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;

import java.time.Instant;
import java.util.List;
import java.util.UUID;

import static ai.suncode.common.exception.ErrorCode.CONFLICT;
import static ai.suncode.common.exception.ErrorCode.HOST_NOT_FOUND;
import static ai.suncode.common.exception.ErrorCode.SESSION_NOT_FOUND;

@RestController
@RequestMapping("/v1/mobile")
@RequiredArgsConstructor
public class MobileRemoteController {
    private final RemoteAuthService authService;
    private final RemoteRelayService relayService;

    @GetMapping("/health")
    public ApiBaseRet<?> health() {
        return ApiBaseRet.success(new HealthData("ok", Instant.now()));
    }

    @PostMapping("/pairings/exchange")
    public ApiBaseRet<?> exchange(@RequestBody PairingExchangeRequest request) {
        String hostId = ServiceContext.current().hostId();
        authService.requirePairingHost(hostId, request == null ? null : request.pairingCode());
        ApiBaseRet<?> response = ApiBaseRet.success(authService.exchange(request));
        relayService.notifyDesktop(hostId, ServiceContext.current().requestId(), "mobile.pairings.exchange.succeeded",
                java.util.Map.of("pathParam", java.util.Map.of(), "queryParam", java.util.Map.of(), "requestBody", java.util.Map.of("deviceName", request.deviceName())));
        return response;
    }

    @PostMapping("/auth/refresh")
    public ApiBaseRet<?> refresh(@RequestBody RefreshTokenRequest request) {
        return ApiBaseRet.success(authService.refreshMobile(ServiceContext.current().hostId(), request));
    }

    @PostMapping("/auth/logout")
    public ResponseEntity<Void> logout() {
        String hostId = ServiceContext.current().hostId();
        authService.logoutToken(ServiceContext.current().token());
        relayService.notifyDesktop(hostId, ServiceContext.current().requestId(), "mobile.auth.logout",
                java.util.Map.of("pathParam", java.util.Map.of(), "queryParam", java.util.Map.of(), "requestBody", java.util.Map.of()));
        return ResponseEntity.noContent().build();
    }

    @GetMapping("/hosts/{hostId}")
    public ApiBaseRet<?> host(@PathVariable String hostId) {
        return ApiBaseRet.success(relayService.hosts().stream().filter(host -> host.id().equals(hostId)).findFirst()
                .orElseThrow(() -> new BusinessException(HOST_NOT_FOUND)));
    }

    @GetMapping("/hosts/{hostId}/projects")
    public ApiBaseRet<?> projects(@PathVariable String hostId) {
        DesktopResponse response = relayService.request(hostId, null, "projects.list", empty(), ServiceContext.current().requestId());
        return responseEnvelope(response, ProjectsData.class);
    }

    @GetMapping("/hosts/{hostId}/sessions")
    public ApiBaseRet<?> sessions(@PathVariable String hostId,
                                  @RequestParam(required = false) String projectId,
                                  @RequestParam(required = false) String cursor,
                                  @RequestParam(required = false) Integer limit) {
        DesktopCommandPayload payload = DesktopCommandPayload.listSessions(projectId, cursor, limit);
        java.util.Map<String, Object> query = new java.util.LinkedHashMap<>();
        if (projectId != null) query.put("projectId", projectId);
        if (cursor != null) query.put("cursor", cursor);
        if (limit != null) query.put("limit", limit);
        return responseEnvelope(relayService.request(hostId, null, "sessions.list", payload, ServiceContext.current().requestId(), query), SessionPageData.class);
    }

    @PostMapping("/hosts/{hostId}/sessions")
    public ResponseEntity<ApiBaseRet<?>> createSession(@PathVariable String hostId,
                                                       @RequestBody CreateSessionRequest request) {
        DesktopResponse response = relayService.request(hostId, null, "session.create", DesktopCommandPayload.createSession(request), ServiceContext.current().requestId());
        String sessionId = sessionId(response.payload());
        relayService.rememberSession(sessionId, hostId);
        return ResponseEntity.status(HttpStatus.CREATED).body(ApiBaseRet.success(new CommandAcceptedData(sessionId)));
    }

    @GetMapping("/hosts/{hostId}/sessions/{sessionId}")
    public ApiBaseRet<?> session(@PathVariable String hostId, @PathVariable String sessionId) {
        return responseEnvelope(relayService.request(hostId, sessionId, "session.get", empty(), ServiceContext.current().requestId()), Object.class);
    }

    @GetMapping(value = "/hosts/{hostId}/sessions/{sessionId}/events", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
    public SseEmitter sessionEvents(@PathVariable String hostId, @PathVariable String sessionId,
                                    @RequestHeader(value = "Last-Event-ID", required = false) String lastEventId) {
        return relayService.connectMobileSession(hostId, sessionId, lastEventId);
    }

    @PostMapping("/hosts/{hostId}/sessions/{sessionId}/messages")
    public ApiBaseRet<?> message(@PathVariable String hostId,
                                 @PathVariable String sessionId,
                                 @RequestBody SendMessageRequest request) {
        return command(hostId, sessionId, "session.message", DesktopCommandPayload.sendMessage(request));
    }

    @PostMapping("/hosts/{hostId}/sessions/{sessionId}/approvals/{approvalId}")
    public ApiBaseRet<?> approval(@PathVariable String hostId,
                                  @PathVariable String sessionId,
                                  @PathVariable String approvalId,
                                  @RequestBody ApprovalResolutionRequest request) {
        return command(hostId, sessionId, "approval.resolve", DesktopCommandPayload.resolveApproval(approvalId, request));
    }

    @PostMapping("/hosts/{hostId}/sessions/{sessionId}/questions/{questionId}/reply")
    public ApiBaseRet<?> question(@PathVariable String hostId,
                                  @PathVariable String sessionId,
                                  @PathVariable String questionId,
                                  @RequestBody QuestionReplyRequest request) {
        return command(hostId, sessionId, "question.reply", DesktopCommandPayload.replyQuestion(questionId, request));
    }

    @PostMapping("/hosts/{hostId}/sessions/{sessionId}/cancel")
    public ApiBaseRet<?> cancel(@PathVariable String hostId,
                                @PathVariable String sessionId) {
        command(hostId, sessionId, "turn.cancel", empty());
        return ApiBaseRet.success();
    }

    @PostMapping("/hosts/{hostId}/sessions/{sessionId}/retry")
    public ApiBaseRet<?> retry(@PathVariable String hostId,
                               @PathVariable String sessionId) {
        command(hostId, sessionId, "turn.retry", empty());
        return ApiBaseRet.success();
    }

    @GetMapping("/hosts/{hostId}/sync")
    public ApiBaseRet<?> sync(@PathVariable String hostId) {
        HostDto host = relayService.hosts().stream().filter(item -> item.id().equals(hostId)).findFirst().orElse(null);
        com.fasterxml.jackson.databind.JsonNode snapshot = relayService.snapshot(hostId);
        List<java.util.Map<String, Object>> sessions = snapshot != null && snapshot.has("sessions")
                ? MarshallingUtils.convertValue(snapshot.get("sessions"), List.class) : List.of();
        return ApiBaseRet.success(new SyncData(UUID.randomUUID().toString(), false, host, sessions, false));
    }

    private ApiBaseRet<?> command(String hostId, String sessionId, String command, DesktopCommandPayload payload) {
        DesktopResponse response = relayService.request(hostId, sessionId, command, payload, ServiceContext.current().requestId());
        return responseEnvelope(response, CommandAcceptedData.class);
    }

    private <T> ApiBaseRet<?> responseEnvelope(DesktopResponse response, Class<T> type) {
        if (!response.success()) {
            throw new BusinessException(CONFLICT, response.message() == null ? "desktop_request_failed" : response.message());
        }
        T data = response.payload() == null ? null : MarshallingUtils.fromJson(response.payload(), type);
        return ApiBaseRet.success(data);
    }

    private String requireHost(String sessionId) {
        String host = relayService.hostForSession(sessionId);
        if (host == null) {
            throw new BusinessException(SESSION_NOT_FOUND);
        }
        return host;
    }

    private static DesktopCommandPayload empty() {
        return new DesktopCommandPayload();
    }

    private static String sessionId(String payload) {
        CommandAcceptedData created = payload == null ? null : MarshallingUtils.fromJson(payload, CommandAcceptedData.class);
        if (created == null || created.sessionId() == null || created.sessionId().isBlank()) {
            return UUID.randomUUID().toString();
        }
        return created.sessionId();
    }

}