# contracts/remote-control/mobile.openapi.yaml
openapi: 3.2.0
info:
  title: SunCode Mobile Remote Control API
  version: 1.0.0
  description: |
    HTTP API between the SunCode CMP Mobile client and the Java Remote Server.
    The Remote Server relays bounded requests to an outbound Desktop connection.
    It is not the Rust agent and never exposes provider wire payloads, secrets, or
    absolute Desktop filesystem paths.
  license:
    name: SunCode project license
    url: https://github.com/suncode-ai/suncode/blob/main/LICENSE
servers:
  - url: https://{remoteServerHost}
    description: Remote Server HTTPS endpoint
    variables:
      remoteServerHost:
        default: relay.suncode.example
...
paths:
...
  /v1/mobile/pairings/exchange:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: exchangePairingCode
      tags: [Pairing]
      security: []
      summary: Exchange a one-time Server-generated Mobile pairing code
      description: |
        The code and X-Host-Id are provided by the Desktop-generated pairing QR
        link. The code was generated by the Remote Server and is bound to the
        Host ID returned by Desktop pairing; Desktop only transports it into the
        QR payload. The Host ID must match the one that issued the one-time code.
        A successful exchange consumes the code atomically and immediately.
        Reuse, expiry, a Host ID mismatch, or a failed Desktop pairing
        transaction never returns a credential. After a successful exchange, the Server notifies the paired Desktop using
        `mobile.pairings.exchange.succeeded`, correlated with this request's
        `X-Request-Id`. The notification never contains the pairing code or
        issued tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingExchangeRequest'
            examples:
              exchangePairingCode:
                summary: Pair a Mobile device with a Desktop
                value:
                  pairingCode: "482917"
                  deviceName: "Long's phone"
      responses:
        '200':
          description: Mobile credential and paired Host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingExchangeResponse'
              examples:
                pairingExchanged:
                  summary: Mobile credentials and paired Host
                  value:
                    code: 0
                    data:
                      accessToken: "opaque-access-token"
                      refreshToken: "opaque-refresh-token"
                      accessTokenExpiresAt: "2026-10-01T03:30:00Z"
                      host:
                        id: "host-01J8YQ6M7A"
                        displayName: "Development PC"
                        connectionState: "connected"
                        agentVersion: "0.1.0"
        '400':
          $ref: '#/components/responses/PlainInvalidRequest'
        '409':
          description: Pairing code was already consumed or pairing is pending elsewhere
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlainApiBaseRetError'
        '410':
          description: Pairing code expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlainApiBaseRetError'
        '503':
          $ref: '#/components/responses/PlainDesktopUnavailable'
  /v1/mobile/auth/refresh:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: refreshAccessToken
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              plaintext:
                value:
                  refreshToken: "opaque-refresh-token"
      responses:
        '200':
          description: Rotated access credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      accessToken: "opaque-access-token"
                      refreshToken: "opaque-refresh-token"
                      accessTokenExpiresAt: "2026-10-01T03:30:00Z"
        '401':
          $ref: '#/components/responses/PlainUnauthorized'
  /v1/mobile/auth/logout:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: logoutCurrentDevice
      tags: [Authentication]
      summary: Log out the current Mobile device
      description: |
        This does not revoke any other Mobile device or Desktop pairing. After
        successfully invalidating the current credential, the Server notifies
        the paired Desktop using `mobile.auth.logout`, correlated with this
        request's `X-Request-Id`. The notification never contains an access or
        refresh token. The current contract has no stable Mobile device ID, so
        Desktop can observe the logout but cannot identify which Mobile device
        logged out when several are paired to the same Host.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Current device credential invalidated
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/mobile/hosts/{hostId}:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getHost
      tags: [Hosts]
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Host detail
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HostResponse'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      id: "host-01J8YQ6M7A"
                      displayName: "Development PC"
                      connectionState: connected
                      lastSeenAt: "2026-10-02T01:47:22.778Z"
                      agentVersion: "0.1.0"
                encrypted:
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/mobile/hosts/{hostId}/projects:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listHostProjects
      tags: [Hosts]
      description: Forwarded to Desktop as the `mobile.projects.list` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Projects exposed by the Host
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetProjects'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      items:
                        - id: "project-01J8YQ6M7C"
                          displayName: "suncode"
                          relativeRoot: "workspace"
                encrypted:
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/mobile/hosts/{hostId}/sessions:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listSessions
      tags: [Sessions]
      summary: List active primary Sessions on one Host
      description: |
        Archived Sessions and child/subagent Sessions are excluded. The default
        response is scoped to the Host identified in the path; Project filtering is optional.
        Forwarded to Desktop as the `mobile.sessions.list` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - $ref: '#/components/parameters/ProjectFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Cursor-paginated active primary Sessions
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetSessionPage'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      items: []
                      nextCursor: null
                      hasMore: false
                encrypted:
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createSession
      tags: [Sessions]
      summary: Create a primary Session on a Host Project
      description: Forwarded to the path Host as the `mobile.sessions.create` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateSessionRequest'
                - $ref: '#/components/schemas/EncryptedPayload'
            examples:
              withImages:
                summary: Create a Session with an initial image
                value:
                  projectId: project-01
                  title: Review screenshot
                  firstMessage: Please inspect this screenshot.
                  images:
                    - mimeType: image/png
                      data: iVBORw0KGgoAAAANSUhEUgAAAAEAAAAB
                      fileName: screenshot.png
              encrypted:
                summary: End-to-end encrypted Session creation request
                value:
                  encPayload: "e2e-v1:base64url-ciphertext"
      responses:
        '201':
          description: Primary Session created. The response contains its ID so Mobile can immediately open its SSE stream.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetSessionCreated'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      sessionId: "session-01J8YQ6M7B"
                encrypted:
                  summary: End-to-end encrypted Session creation response
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/DesktopUnavailable'
  /v1/mobile/hosts/{hostId}/sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getSession
      tags: [Sessions]
      description: Forwarded to Desktop as the `mobile.sessions.get` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Session snapshot and current pending interaction
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetSessionDetail'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      id: "session-01J8YQ6M7B"
                      title: "Investigate remote control"
                      kind: primary
                      project:
                        id: "project-01J8YQ6M7C"
                        displayName: "suncode"
                      state: idle
                      updatedAt: "2026-10-02T01:45:10Z"
                      preview: "Review the remote control contract"
                      revision: 7
                      eventSequence: 42
                      archived: false
                      messages: []
                      pendingApproval: null
                      pendingQuestion: null
                encrypted:
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Session was archived elsewhere and is no longer in the mobile active projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseRetError'
  /v1/mobile/hosts/{hostId}/sessions/{sessionId}/events:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/LastEventId'
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: streamSessionEvents
      tags: [Sessions]
      summary: Stream one Session using Server-Sent Events
      description: |
        Opens a server-to-Mobile SSE stream for exactly one active primary
        Session. Without `Last-Event-ID`, the first application event is
        `session.snapshot`, followed by live `agent.event` frames. The snapshot
        is fetched from Desktop using the `mobile.sessions.events.open` SSE
        event, correlated to this request's `X-Request-Id`. The snapshot
        has its own stable SSE `id` and sequence. On a
        reconnect with a retained event ID, the server replays only events with
        a greater sequence before live delivery. The stream has no command
        channel. Heartbeats are SSE comments and do not advance the cursor.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: UTF-8 text/event-stream for the selected Session
          headers:
            Cache-Control:
              description: Must include no-cache for a live stream.
              schema: { type: string }
            X-Session-Sequence:
              description: Current per-Session sequence at stream establishment.
              schema: { type: integer, minimum: 0 }
          content:
            text/event-stream:
              schema:
                type: array
                description: Sequential SSE stream represented as its parsed event items.
                items:
                  $ref: '#/components/schemas/SessionSseStreamItem'
              itemSchema:
                $ref: '#/components/schemas/SessionSseStreamItem'
              examples:
                snapshot:
                  summary: Initial Session snapshot
                  value:
                    - event: session.snapshot
                      id: host-01:42
                      data: '{"event_id":"host-01:42","session_id":"session-01","sequence":42,"session_revision":7,"snapshot":{"id":"session-01","title":"Investigate remote control","kind":"primary","project":{"id":"project-01","displayName":"suncode"},"state":"idle","updatedAt":"2026-10-01T02:30:00Z","preview":"Review the remote control contract","revision":7,"eventSequence":42,"archived":false,"messages":[],"pendingApproval":null,"pendingQuestion":null}}'
                agentEvent:
                  summary: Normalized Agent event
                  value:
                    - event: agent.event
                      id: host-01:43
                      data: '{"event_id":"host-01:43","sequence":43,"session_revision":8,"session_id":"session-01","occurred_at":"2026-10-01T02:31:00Z","event_type":"message.assistant","payload":{"message_id":"message-01","turn_id":"turn-01","call_id":null,"message":{"role":"assistant","content":[{"type":"text","text":"The contract is consistent."}]},"usage":{"input_tokens":120,"output_tokens":18,"total_tokens":138},"finish_reason":"stop"}}'
                encryptedEvent:
                  summary: End-to-end encrypted Session event
                  value:
                    - event: agent.event
                      id: host-01:44
                      data: '{"encPayload":"e2e-v1:base64url-ciphertext"}'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Last-Event-ID is outside the retained replay window; fetch a fresh snapshot and reconnect without the cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseRetError'
        '429':
          description: Stream admission is rate limited; retry after the indicated delay.
          headers:
            Retry-After:
              schema: { type: integer, minimum: 1 }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseRetError'
        '503':
          $ref: '#/components/responses/DesktopUnavailable'
  /v1/mobile/hosts/{hostId}/sessions/{sessionId}/messages:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: sendSessionMessage
      tags: [Sessions]
      summary: Send one user message to a primary Session
      description: Forwarded to Desktop as the `mobile.sessions.messages.send` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/SendMessageRequest'
                - $ref: '#/components/schemas/EncryptedPayload'
            examples:
              withImages:
                summary: Send a message with an image
                value:
                  text: Continue from this screenshot.
                  images:
                    - mimeType: image/jpeg
                      data: /9j/4AAQSkZJRgABAQAAAQABAAD
                      fileName: context.jpg
              encrypted:
                summary: End-to-end encrypted message request
                value:
                  encPayload: "e2e-v1:base64url-ciphertext"
      responses:
        '202':
          description: Message accepted for Desktop delivery
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetCommandAccepted'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      sessionId: "session-01J8YQ6M7B"
                encrypted:
                  summary: End-to-end encrypted command response
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/DesktopUnavailable'
  /v1/mobile/hosts/{hostId}/sessions/{sessionId}/approvals/{approvalId}:
    parameters:
      - $ref: '#/components/parameters/HostHeader'
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: resolveApproval
      tags: [Sessions]
      summary: Resolve a pending Rust approval
      description: Forwarded to Desktop as the `mobile.sessions.approvals.resolve` SSE event.
      security:
        - bearerAuth: []
      parameters:
        - name: hostId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - name: approvalId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ApprovalResolutionRequest'
                - $ref: '#/components/schemas/EncryptedPayload'
            examples:
              plaintext:
                value:
                  action: allow_once
                  expectedRevision: 7
              encrypted:
                summary: End-to-end encrypted approval request
                value:
                  encPayload: "e2e-v1:base64url-ciphertext"
      responses:
        '202':
          description: Approval resolution accepted
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiBaseRetCommandAccepted'
                  - $ref: '#/components/schemas/EncryptedPayload'
              examples:
                plaintext:
                  value:
                    code: 0
                    data:
                      sessionId: "session-01J8YQ6M7B"
                encrypted:
                  summary: End-to-end encrypted approval response
                  value:
                    encPayload: "e2e-v1:base64url-ciphertext"
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Stale revision or approval already resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBaseRetError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    HostHeader:
      name: X-Host-Id
      in: header
      required: true
      description: Host identity associated with this Mobile request.
      schema: { type: string, minLength: 1, maxLength: 200 }
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      description: Unique request ID used for correlation and mutation deduplication. Reuse it when retrying the same logical request.
      schema: { type: string, minLength: 1, maxLength: 200 }
...
  responses:
...
    PlainInvalidRequest:
      description: Request validation failed; pairing and token refresh bodies are plaintext.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PlainApiBaseRetError' }
...
    PlainDesktopUnavailable:
      description: Selected Desktop is unavailable; pairing responses are plaintext.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PlainApiBaseRetError' }
  schemas:
...
    PlainApiBaseRetError:
      $ref: '#/components/schemas/ApiBaseRet'
...
    PairingExchangeRequest:
      type: object
      required: [pairingCode, deviceName]
      additionalProperties: false
      properties:
        pairingCode:
          type: string
          minLength: 1
          description: One-time code generated by the Remote Server and bound to the Host. Desktop only embeds it in the QR payload; the Server consumes it atomically on the first successful exchange.
        deviceName:
          type: string
          minLength: 1
          maxLength: 120
    PairingExchangeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiBaseRet'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/PairingExchangeData' }
...